Revert renaming spring module to spring-boot-3
