+++ README.md
@@ -16,7 +16,7 @@ Server-side <a href="https://inertiajs.com/">Inertia.js</a> adapter for JVM back
 
 Inertia provides a protocol to render components written in modern frontend frameworks such as React, Vue.js, and Angular, directly from the backend. Inertia4J is a JVM adapter implementing such a protocol, and currently offers first-class support for Spring and Ktor. Below, you can find usage instructions for each framework.
 
-* [Using with Spring Boot 3 (Java / Kotlin)](/inertia4j.spring-boot-3/README.md)
+* [Using with Spring Boot 3 (Java / Kotlin)](/inertia4j.spring/README.md)
 * [Using with Spring Boot 4 (Java / Kotlin)](/inertia4j.spring-boot-4/README.md)
 * [Using with Ktor (Kotlin)](/inertia4j.ktor/README.md)
 * [Advanced Usage and Extending Inertia4J](#advanced-usage-and-extending-inertia4j)
