+++ inertia4j.spring/src/main/java/io/github/inertia4j/spring/InertiaHttpServletRequest.java
@@ -1,4 +1,4 @@
-package io.github.inertia4j.springboot3;
+package io.github.inertia4j.spring;
 
 import io.github.inertia4j.core.HttpRequest;
 import jakarta.servlet.http.HttpServletRequest;
