+++ inertia4j.spring/src/main/java/io/github/inertia4j/spring/VersionProvider.java
@@ -1,4 +1,4 @@
-package io.github.inertia4j.springboot3;
+package io.github.inertia4j.spring;
 
 /**
  * Functional interface for providing the current Inertia asset version.
