+++ inertia4j.spring/src/test/java/io/github/inertia4j/spring/InertiaMockMvcTest.java
@@ -1,4 +1,4 @@
-package io.github.inertia4j.springboot3;
+package io.github.inertia4j.spring;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
