+++ inertia4j.spring/src/test/resources/templates/app.html
